==> 01-Data.ipynb <==
import pandas as pd
import ijson
import glob

res = []
for path in glob.glob("data/args.me/*.json"):
    with open(path) as f:
        objects = ijson.items(f, 'arguments.item')
        for x in objects:
            res.append({"id": x["id"], "premise": " ".join(y["text"] for y in x["premises"]), "conclusion": x["conclusion"]})

res = pd.DataFrame(res)
res.to_parquet("data/args.parquet")
