==> 02-Quality.ipynb <==
# Data Import

import pandas as pd

df = pd.read_parquet("data/args.parquet").drop("conclusion", axis=1)

IDS = df["id"].tolist()
SAMPLES = df["premise"].tolist()

del(df)

df = pd.read_csv("data/webis-argquality20-full.csv")

# Vectorization

from sklearn.feature_extraction.text import TfidfVectorizer

vectorizer = TfidfVectorizer(
    lowercase=True, 
    stop_words="english",
    max_features = 2**18
)
SAMPLES = vectorizer.fit_transform(SAMPLES)

# Support Vector Machine for Argument Classification

from sklearn.svm import SVC
from sklearn.model_selection import cross_val_score

X = df["Premise"].astype(str).tolist()
Y = (df["Combined Quality"].astype(float) != -4.0).tolist()

X = vectorizer.transform(X)

classificator = SVC()
classificator.fit(X,Y)

scores = cross_val_score(classificator, X, Y, cv=10, scoring='f1')
"F1-score: {}".format(scores.mean().round(4))

# Support Vector Regression for Quality Prediction

from sklearn.svm import LinearSVR
from sklearn.preprocessing import MinMaxScaler
from sklearn.model_selection import cross_val_score

X = df.loc[df["Combined Quality"] != -4.0, "Premise"].astype(str).values
Y = df.loc[df["Combined Quality"] != -4.0, "Combined Quality"].astype(float).values

scaler = MinMaxScaler()
Y = scaler.fit_transform(Y.reshape(-1, 1)).ravel()
X = vectorizer.transform(X).todense()

regressor = LinearSVR(max_iter=10_000)
regressor.fit(X, Y)

scores = cross_val_score(regressor, X, Y, cv=10)
"MSE: {}".format(scores.mean().round(4))

# Predicting Scores for the args.me dataset

PREDICTIONS = classificator.predict(SAMPLES)
QUALITY = regressor.predict(SAMPLES)

df = pd.DataFrame(zip(IDS, PREDICTIONS, QUALITY), columns=["id", "is_argument", "quality"])
df.loc[(df.is_argument == False),"quality"] = 0
df = df.drop("is_argument", axis=1)
df.to_parquet("data/quality.parquet")
