==> 03-Clustering.ipynb <==
# Data Import

import pandas as pd
df = pd.read_parquet("data/args.parquet")

# Embeddings

import pandas as pd
import tensorflow_hub as hub
embed = hub.load("https://tfhub.dev/google/universal-sentence-encoder/4")

df["conclusion_embedded"] = df["conclusion"].apply(lambda x: embed([x]).numpy()[0])

del(embed)

df.to_parquet("data/embeddings.parquet")

# Clustering

import numpy as np
import pandas as pd

df = pd.read_parquet("data/embeddings.parquet")

ID = np.vstack(df["id"].values)
X = np.vstack(df["conclusion_embedded"].values)

from cuml.cluster import KMeans

clusterer = KMeans(n_clusters=300)
clusterer.fit(X)
clusters = clusterer.labels_

df = pd.DataFrame(zip(ID,X,clusterer.labels_), columns=["id","embedding","cluster"])

centroids = (
    df
    .loc[:,["embedding","cluster"]]
    .groupby("cluster")
    .apply(lambda group: np.vstack(group.values).mean(axis=0)[0])
    .reset_index()
    .rename({"cluster":"cluster", 0:"centroid"}, axis=1)
)

df["id"] = df["id"].apply(lambda x: x[0])
df = df.merge(centroids, on="cluster", how="left").drop("embedding",axis=1)

df.to_parquet("data/clusters.parquet")
