==> 04-Consolidate.ipynb <==
import pandas as pd

df_text = pd.read_parquet("data/args.parquet")
df_quality = pd.read_parquet("data/quality.parquet")
df_cluster = pd.read_parquet("data/clusters.parquet")

df = (
    pd.merge(
        df_text,
        (
            pd.merge(
                df_cluster,
                df_quality,
                on = "id",
                how = "inner"
            )
        ),
        on = "id",
        how = "inner"
    )
    .sort_values(["quality"], ascending=False)
    .drop_duplicates(subset="id", keep="first")
    .sort_values(["cluster","quality"], ascending=False)
    .reset_index(drop=True)
)

df["quality"] = (df["quality"] - df["quality"].min())/(df["quality"].max() - df["quality"].min())

df.to_parquet("data/index.parquet")
